FAC-762: Rename GenerateCodeUtil to GenerateHashCodeUtil
