+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/application/service/answerrange/CreateAnswerRangeService.java
@@ -10,7 +10,7 @@ import org.flickit.assessment.kit.application.port.out.kitversion.LoadKitVersion
 import org.springframework.stereotype.Service;
 
 import static org.flickit.assessment.common.error.ErrorMessageKey.COMMON_CURRENT_USER_NOT_ALLOWED;
-import static org.flickit.assessment.common.util.GenerateCodeUtil.generateCode;
+import static org.flickit.assessment.common.util.GenerateHashCodeUtil.generateCode;
 
 @Service
 @Transactional
