+++ flickit-assessment-kit/src/main/java/org/flickit/assessment/kit/application/service/questionnaire/UpdateQuestionnaireService.java
@@ -14,7 +14,7 @@ import java.time.LocalDateTime;
 import java.util.UUID;
 
 import static org.flickit.assessment.common.error.ErrorMessageKey.COMMON_CURRENT_USER_NOT_ALLOWED;
-import static org.flickit.assessment.common.util.GenerateCodeUtil.generateCode;
+import static org.flickit.assessment.common.util.GenerateHashCodeUtil.generateCode;
 
 @Service
 @Transactional
